Tabs: The political affiliation tab now works. WIP: upvoted and comments tab
