+++ app/src/main/java/me/caelumterrae/fbunewsapp/fragments/UserFragmentTabs/CommentFragment.java
@@ -1,0 +1,31 @@
+package me.caelumterrae.fbunewsapp.fragments.UserFragmentTabs;
+
+import android.content.Context;
+import android.net.Uri;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import me.caelumterrae.fbunewsapp.R;
+
+public class CommentFragment extends Fragment {
+    public CommentFragment(){
+
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_comment, container, false);
+    }
+
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+
+    }
+}
