+++ app/src/main/java/me/caelumterrae/fbunewsapp/fragments/UserFragmentTabs/UpvotedFragment.java
@@ -1,0 +1,32 @@
+package me.caelumterrae.fbunewsapp.fragments.UserFragmentTabs;
+
+import android.content.Context;
+import android.net.Uri;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import me.caelumterrae.fbunewsapp.R;
+
+
+public class UpvotedFragment extends Fragment {
+    public UpvotedFragment(){
+
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_upvoted, container, false);
+    }
+
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+
+    }
+}
